requirements.txt
boto==2.38.0
click==5.1
ecdsa==0.13
Fabric==1.10.2
paramiko==1.15.2
pycrypto==2.6.1
python-dotenv==0.1.3
six==1.9.0
wheel==0.24.0
